Service-Layer Architecture & Decoupling Overhaul
